pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pop-agent"
version = "0.1.0"
description = "Multi-agent popular science generation CLI and API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115",
  "httpx>=0.27",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "rich>=13.7",
  "typer>=0.12",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
]

[project.scripts]
pop-agent = "pop_agent.cli:app"

[tool.setuptools.packages.find]
include = ["pop_agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
